pyproject.toml
[project]
name = "whoop-mcp-server"
version = "0.1.0"
description = "MCP server for Whoop API integration"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.2,<2",
    "httpx>=0.27,<1",
    "python-dotenv>=1.0,<2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
whoop-mcp-server = "whoop_server:main"

[tool.hatch.build.targets.wheel]
packages = ["src"] 